requirements.txt
astroid==2.5.2
atomicwrites==1.4.0
attrs==19.3.0
certifi==2019.9.11
chardet==3.0.4
charset-normalizer==2.0.4
colorama==0.4.4
# Django==2.2.6
django-debug-toolbar==3.2.2
django-debug-toolbar-force==0.1.8
django-filter==2.4.0
django-nine==0.2.4
# djangorestframework==3.12.4
# djangorestframework-simplejwt==4.7.2
Faker==8.1.2
flake8==3.9.2
idna==2.8
importlib-metadata==1.5.0
iniconfig==1.1.1
lazy-object-proxy==1.6.0
lxml==4.6.3
mccabe==0.6.1
mixer==7.1.2
more-itertools==8.2.0
mypy==0.812
mypy-extensions==0.4.3
packaging==20.1
pep8-naming==0.11.1
Pillow==8.0.0
pluggy==0.13.1
# psycopg2-binary==2.8.6
py==1.8.1
pycodestyle==2.7.0
pyflakes==2.3.1
# PyJWT==2.1.0
pylint==2.7.4
pyparsing==2.4.6
pytest==5.3.5
pytest-django==3.8.0
pytest-pythonpath==0.7.3
python-dateutil==2.8.1
# pytz==2019.3
requests==2.22.0
rope==0.18.0
six==1.14.0
sorl-thumbnail==12.6.3
# sqlparse==0.3.0
testfixtures==6.17.1
text-unidecode==1.3
toml==0.10.2
typed-ast==1.4.2
typing-extensions==3.7.4.3
urllib3==1.25.6
wcwidth==0.1.8
wrapt==1.12.1
yapf==0.31.0
zipp==2.2.0

asgiref==3.2.10
Django==3.0.5
djangorestframework==3.11.0
djangorestframework-simplejwt==4.3.0
# gunicorn==20.0.4
psycopg2-binary==2.8.6
PyJWT==1.7.1
pytz==2020.1
sqlparse==0.3.1 